chore(email): folder restructure
