+++ apps/admin/app/(dashboard)/(modules)/email/settings/page.tsx
@@ -1,4 +1,4 @@
-import { Settings } from '@/components/email/settings';
+import { Settings } from '@/components/email/settings/settings';
 import { getEmailSettings } from '@/lib/api/email';
 
 export default async function EmailSettings() {
